pyproject.toml
[project]
name = "projectgenius"
version = "0.1.0"
description = "AI-powered education platform built with Flask"
authors = [
    {name = "ProjectGenius Team", email = "team@projectgenius.edu"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "flask-migrate>=4.0.5",
    "flask-wtf>=1.2.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "werkzeug>=3.1.3",
    "redis>=5.0.1",
    "celery>=5.3.6",
    "pillow>=10.1.0",
    "pyjwt>=2.8.0",
    "bcrypt>=4.1.2",
    "markdown>=3.5.1",
    "bleach>=6.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.1",
    "isort>=5.12.0",
    "pre-commit>=3.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=projectgenius"
